added filtering system for pokemon singles

+++ FuwaCards/Models/PokemonSingles.cs
@@ -6,10 +6,11 @@
 		public string Name { get; set; }
 		public string Rarity { get; set; }
 		public string Set { get; set; }	
-		public double Price { get; set; }
+		public decimal Price { get; set; }
 		public string Condition { get; set; }
 		public string Image { get; set; }
 		public string AltTex { get; set; }
 		public int Quantity { get; set; }
+
 	}
 }
